modified controllers and also added a otpHeper file
